app/requirements.txt
fastapi
uvicorn[standard]
asyncpg
python-dotenv
loguru
paho-mqtt
websockets
redis[hiredis]